Move non site related code from SiteUtils.java to DateUtils class
